+++ test/TestableHttpClient.IntegrationTests/UsingIHttpClientFactory.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Builder;
+﻿#if NETCOREAPP || NET
+using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.TestHost;
@@ -115,3 +116,4 @@ public class UsingIHttpClientFactory
         }
     }
 }
+#endif